// resources/js/apps/admin/components/spells/SpellForm/types.ts
export type SpellErrors = {
  name?: string | string[]
  range?: string | string[]
  components?: string | string[]
  materials?: string | string[]
  ritual?: string | string[]
  concentration?: string | string[]
  duration?: string | string[]
  casting_time?: string | string[]
  level?: string | string[]
  school?: string | string[]
  description?: string | string[]
  higher_levels?: string | string[]
}